requirements.txt
scikit-learn
numpy
python-dotenv
matplotlib
torch==1.11.0
torchvision==0.12.0